homework-10: fixed
исправления
